+++ src/api_server/impl/SMFRegistrationsCollectionApiImpl.cpp
@@ -58,7 +58,7 @@ void SMFRegistrationsCollectionApiImpl::query_smf_reg_list(
   m_udr_app->handle_query_smf_reg_list(ueId, response_data, http_code);
 
   code = static_cast<Pistache::Http::Code>(http_code);
-  Logger::udr_server().debug("HTTP Response code %d.\n", code);
+  Logger::udr_server().debug("HTTP Response code %d.\n", (int) code);
   // content type
   response.headers().add<Pistache::Http::Header::ContentType>(
       Pistache::Http::Mime::MediaType("application/json"));
